hw10 edit after code-review2
