+++ hw10/src/main/java/ru/otus/hw/controller/ErrorControllerAdvice.java
@@ -5,12 +5,13 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import ru.otus.hw.exceptions.EntityNotFoundException;
 
 @ControllerAdvice
 public class ErrorControllerAdvice {
-    @ExceptionHandler(NotFoundException.class)
+    @ExceptionHandler(EntityNotFoundException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public String errorPage(NotFoundException ex, Model model) {
+    public String errorPage(EntityNotFoundException ex, Model model) {
         model.addAttribute("message", ex.getMessage());
         return "error";
     }
