+++ hw10/src/main/java/ru/otus/hw/exceptions/EntityNotFoundException.java
@@ -1,5 +1,8 @@
 package ru.otus.hw.exceptions;
 
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
 public class EntityNotFoundException extends RuntimeException {
     public EntityNotFoundException(String message) {
         super(message);
