+++ hw10/src/main/resources/templates/book/delete.html
@@ -35,10 +35,7 @@
                     'Accept': 'application/json',
                     'Content-Type': 'application/json'
                 }
-            }).then(window.location.href = "/")
-            //почему после первого вызова когда переходим на страницу со списком, видим там все книги?
-            // если обновить страницу, то удаленная книга уходит. При последующих попытках удаления
-            // все происходит как задумано. Только самое первое нажатие тупит.
+            }).then(response => window.location.href = "/")
         }
 
         function getBook() {
