+++ hw10/src/test/java/ru/otus/hw/controller/PageBookControllerTest.java
@@ -8,12 +8,11 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
-@WebMvcTest({PageController.class})
-public class PageControllerTest {
+@WebMvcTest({PageBookController.class})
+public class PageBookControllerTest {
 
     @Autowired
     private MockMvc mvc;
